refactor textfield
