+++ src/components/atoms/textfield/textfield.tsx
@@ -1,16 +1,10 @@
-import React, { FC } from 'react';
-import {
-  CalendarIconComponent,
-  LocationIconComponent,
-  TimeIconComponent,
-  UserIconComponent
-} from '../icons/index';
+import React, {FC} from 'react';
 
-export type WriteTextfieldProps = {
+export type TextfieldProps = {
   placeholder: string;
 };
 
-const WriteTextfield: FC<WriteTextfieldProps> = ({ placeholder }) => {
+const Textfield: FC<TextfieldProps> = ({ placeholder }) => {
   return <textarea className={
     'bg-slate-100 h-160 w-355 border-1 hover:border-2 border-slate-200 ' +
     'hover:border-slate-300 rounded-s focus:outline focus:outline-none focus:border-2 ' +
@@ -18,4 +12,4 @@ const WriteTextfield: FC<WriteTextfieldProps> = ({ placeholder }) => {
   } placeholder={placeholder}></textarea>;
 };
 
-export default WriteTextfield;
+export default Textfield;
